Get git-server ssh key before any Plastic->Git activity, instead of during startup
During startup there is a risk that the git-server container is also starting/restarting. When that happens, ssh-keyscan will not be able to talk to the SSH server on git-server and therefore it won't get any key. Simplest solution is to repeat key scanning before every command that will talk to Git.