configure.sh will trigger establishing of git-server host keys from within plastic container
Triggering host key setup from configure.sh ensures that host keys are not changed automatically (so no security downside). It also means that both containers are running so ssh-keyscan will be sure to find the git-server container.

There might be a race condition lurking here; what if it takes a long time for git-server to start up? Perhaps there's a need to introduce a check to configure.sh which waits until the git-server SSH server accepts incoming connections, before moving on further.